0x02 functions advance 16  8nd trail

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
  */
 int main(void)
 {
-	long long int pri = 1, sec = 2, sum = 0;
+	unsigned long int pri = 1, sec = 2, sum = 0;
 	int limit = 3;
 
 	printf("1, 2");
 	while (limit <= 98)
 	{
 		sum = pri + sec;
-		printf(", %lli", sum);
+		printf(", %llu", sum);
 		pri = sec;
 		sec = sum;
 		limit++;
